requirements.txt
langgraph==0.5.1
langchain==0.3.26
langchain-openai==0.3.27
langchain-community==0.3.27
pydantic==2.11.7
python-dotenv==1.1.1
graphviz==0.21
